feat: preview window hover actions, with time and action customizations
closes #27 and #32
