+++ DockDoor/Views/Hover Window/WindowPreview.swift
@@ -7,7 +7,6 @@
 
 import SwiftUI
 import Defaults
-import FluidGradient
 
 struct WindowPreview: View {
     let windowInfo: WindowInfo
@@ -17,66 +16,61 @@ struct WindowPreview: View {
     let maxWindowDimension: CGPoint
     let bestGuessMonitor: NSScreen
     let uniformCardRadius: Bool
-    
+
     @Default(.windowTitlePosition) var windowTitlePosition
     @Default(.showWindowTitle) var showWindowTitle
     @Default(.windowTitleDisplayCondition) var windowTitleDisplayCondition
     @Default(.trafficLightButtonsVisibility) var trafficLightButtonsVisibility
     
+    // preview popup action handlers
+    @Default(.tapEquivalentInterval) var tapEquivalentInterval
+    @Default(.previewHoverAction) var previewHoverAction
+
     @State private var isHoveringOverDockPeekPreview = false
     @State private var isHoveringOverWindowSwitcherPreview = false
-    
-    private var calculatedMaxDimensions: CGSize? {
+    @State private var fullPreviewTimer: Timer?
+
+    private var calculatedMaxDimensions: CGSize {
         CGSize(width: self.bestGuessMonitor.frame.width * 0.75, height: self.bestGuessMonitor.frame.height * 0.75)
     }
-    
+
     var calculatedSize: CGSize {
         guard let cgImage = windowInfo.image else { return .zero }
-        
+
         let cgSize = CGSize(width: cgImage.width, height: cgImage.height)
         let aspectRatio = cgSize.width / cgSize.height
         let maxAllowedWidth = maxWindowDimension.x
         let maxAllowedHeight = maxWindowDimension.y
-        
+
         var targetWidth = maxAllowedWidth
         var targetHeight = targetWidth / aspectRatio
-        
+
         if targetHeight > maxAllowedHeight {
             targetHeight = maxAllowedHeight
             targetWidth = aspectRatio * targetHeight
         }
-        
+
         return CGSize(width: targetWidth, height: targetHeight)
     }
-    
-    private func fluidGradient() -> some View {
-        FluidGradient(
-            blobs: [.purple, .blue, .green, .yellow, .red, .purple].shuffled(),
-            highlights: [.red, .orange, .pink, .blue, .purple].shuffled(),
-            speed: 0.45,
-            blur: 0.75
-        )
-        .opacity(0.125)
-    }
-    
+
     private func windowContent(isMinimized: Bool, isHidden: Bool, isSelected: Bool) -> some View {
         Group {
             if isMinimized || isHidden {
                 let width = maxWindowDimension.x > 300 ? maxWindowDimension.x : 300
                 let labelText = isMinimized ? "Minimized" : "Hidden"
-                
+
                 HStack(spacing: 16) {
                     Image(systemName: "eye.slash.fill")
                         .font(.system(size: 20))
                         .foregroundColor(.secondary)
-                    
+
                     Divider()
-                    
+
                     VStack(alignment: .leading) {
                         Text(labelText)
                             .font(.caption)
                             .foregroundStyle(.secondary)
-                        
+
                         TheMarquee(width: width - 100, secsBeforeLooping: 2, speedPtsPerSec: 30, nonMovingAlignment: .leading) {
                             Text(windowInfo.windowName ?? "\(labelText) window")
                                 .font(.system(size: 12, weight: .medium))
@@ -88,22 +82,22 @@ struct WindowPreview: View {
                 .padding()
                 .frame(width: width)
                 .frame(height: 60)
-                .overlay { if isSelected { fluidGradient() }}
+                .overlay { if isSelected { fluidGradient().opacity(0.125) }}
             } else if let cgImage = windowInfo.image {
                 let image = Image(decorative: cgImage, scale: 1.0).resizable().aspectRatio(contentMode: .fill)
-                image.overlay(!isSelected ? nil : fluidGradient().mask(image))
+                image.overlay(!isSelected ? nil : fluidGradient().opacity(0.125).mask(image))
             }
         }
         .frame(width: isMinimized || isHidden ? nil : calculatedSize.width,
                height: isMinimized || isHidden ? nil : calculatedSize.height,
                alignment: .center)
-        .frame(maxWidth: calculatedMaxDimensions?.width, maxHeight: calculatedMaxDimensions?.height)
+        .frame(maxWidth: calculatedMaxDimensions.width, maxHeight: calculatedMaxDimensions.height)
     }
-    
+
     var body: some View {
         let isHighlightedInWindowSwitcher = (index == ScreenCenteredFloatingWindow.shared.currIndex && ScreenCenteredFloatingWindow.shared.windowSwitcherActive)
         let selected = isHoveringOverDockPeekPreview || isHighlightedInWindowSwitcher
-        
+
         ZStack(alignment: .topTrailing) {
             VStack(spacing: 0) {
                 windowContent(isMinimized: windowInfo.isMinimized, isHidden: windowInfo.isHidden, isSelected: selected)
@@ -144,6 +138,7 @@ struct WindowPreview: View {
             withAnimation(.snappy(duration: 0.175)) {
                 if (!ScreenCenteredFloatingWindow.shared.windowSwitcherActive) {
                     isHoveringOverDockPeekPreview = over
+                    handleFullPreviewHover(isHovering: over, action: previewHoverAction)
                 } else {
                     isHoveringOverWindowSwitcherPreview = over
                 }
@@ -151,6 +146,60 @@ struct WindowPreview: View {
         }
         .onTapGesture {
             handleWindowTap()
+        }
+    }
+
+    private func handleFullPreviewHover(isHovering: Bool, action: HoverTimerActions) {
+        if isHovering && !ScreenCenteredFloatingWindow.shared.windowSwitcherActive {
+            switch action {
+            case .none:
+                // Do nothing for .none
+                break
+                
+            case .tap:
+                // If the interval is 0, immediately trigger the tap action
+                if tapEquivalentInterval == 0 {
+                    handleWindowTap()
+                } else {
+                    // Set a timer to trigger the tap action after the specified interval
+                    fullPreviewTimer = Timer.scheduledTimer(withTimeInterval: tapEquivalentInterval, repeats: false) { [self] _ in
+                        DispatchQueue.main.async {
+                            handleWindowTap()
+                        }
+                    }
+                }
+                
+            case .previewFullSize:
+                // If the interval is 0, show the full window preview immediately
+                if tapEquivalentInterval == 0 {
+                    DispatchQueue.main.async {
+                        SharedPreviewWindowCoordinator.shared.showWindow(
+                            appName: windowInfo.appName,
+                            windows: [windowInfo],
+                            mouseScreen: bestGuessMonitor,
+                            overrideDelay: true,
+                            centeredHoverWindowState: .fullWindowPreview
+                        )
+                    }
+                } else {
+                    // If the interval is greater than 0, set a timer to show the full window preview after the specified interval
+                    fullPreviewTimer = Timer.scheduledTimer(withTimeInterval: tapEquivalentInterval, repeats: false) { [self] _ in
+                        DispatchQueue.main.async {
+                            SharedPreviewWindowCoordinator.shared.showWindow(
+                                appName: windowInfo.appName,
+                                windows: [windowInfo],
+                                mouseScreen: bestGuessMonitor,
+                                overrideDelay: true,
+                                centeredHoverWindowState: .fullWindowPreview
+                            )
+                        }
+                    }
+                }
+            }
+        } else {
+            // If the hover state is not active or the window switcher is active, invalidate the timer
+            fullPreviewTimer?.invalidate()
+            fullPreviewTimer = nil
         }
     }
     
@@ -164,14 +213,14 @@ struct WindowPreview: View {
         }
         onTap?()
     }
-    
+
     @ViewBuilder
     private func windowTitleOverlay(selected: Bool) -> some View {
         if selected, let windowTitle = windowInfo.window?.title, !windowTitle.isEmpty, windowTitle != windowInfo.appName {
             let maxLabelWidth = calculatedSize.width - 50
             let stringMeasurementWidth = measureString(windowTitle, fontSize: 12).width + 5
             let width = maxLabelWidth > stringMeasurementWidth ? stringMeasurementWidth : maxLabelWidth
-            
+
             TheMarquee(width: width, secsBeforeLooping: 1, speedPtsPerSec: 20, nonMovingAlignment: .leading) {
                 Text(windowInfo.windowName ?? "Hidden window")
                     .font(.system(size: 12, weight: .medium))
